Rename Camera_manager to CameraOffset, remove singleton & format
