+++ Assets/Prefabs/Main Camera.prefab
@@ -108,10 +108,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   player: {fileID: 328789305087572588, guid: 6a120e85c4b8215448cb02a44cba1de9, type: 3}
-  cameraOffsetXYZ:
-  - 4
-  - 5.6
-  - 4
+  offset: {x: 4, y: 5.6, z: 4}
 --- !u!1 &5254030397172625456
 GameObject:
   m_ObjectHideFlags: 0
@@ -218,10 +215,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   player: {fileID: 328789305087572588, guid: 6a120e85c4b8215448cb02a44cba1de9, type: 3}
-  cameraOffsetXYZ:
-  - 4
-  - 5.6
-  - 4
+  offset: {x: 0, y: 0, z: 0}
 --- !u!1 &6139500068076424284
 GameObject:
   m_ObjectHideFlags: 0
@@ -328,7 +322,4 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   player: {fileID: 328789305087572588, guid: 6a120e85c4b8215448cb02a44cba1de9, type: 3}
-  cameraOffsetXYZ:
-  - 4
-  - 5.6
-  - 4
+  offset: {x: 0, y: 0, z: 0}
